reservation termine

+++ src/app/list-reservation/list-reservation.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Reservation } from '../models/reservation';
 import { ReservationService } from '../reservation.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-reservation',
@@ -26,10 +27,27 @@ listReservation: Reservation[]=[]
     )
   }
   deleteReservation(id:number){
+  Swal.fire({
+    title: 'Etes vous sur ?',
+    text: "Vous ne pourrez pas revenir en arrière!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Oui, Supprimer!'
+  }).then((result)=> {
+  if(result.value){
     this.reservationservice.delete(id).subscribe(
-      data=>(
-        console.log(data)
-      )
-    )
-  }
-}
+      data=>{
+ if(data){
+          Swal.fire(
+            'Reservation supprimé!',
+          ).then(()=>
+           window.location.href = "http://localhost:4200/reservation"
+          )
+        }
+      }
+    )}
+ 
+  })
+}}
